+++ src/shared/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useOutsideClick } from './useOutsideClick'
 export { default as useSelect } from './useSelect'
 export { default as useAutocomplete } from './useAutoComplete'
 export { default as useDebounce } from './useDebounce'
+export { default as useKeyboardSelect } from './useKeyboardSelect'
